backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
joblib>=1.4,<2.0
numpy>=1.26,<3.0
pydantic>=2.7,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
